# Fix performance of DPF vector