closing #2201 

DPF vector became slow because they were always checking if the modification of the vector by
1. making a deep copy of the vector
2. comparing the whole vector

**Why performance changed between 0.13.6 and master?**
Before 0.13.7.dev, it was hard coded that in process dpf vector wouldn't "check modification" and though bypassed dpf server logic. 0.13.7.dev is now relying on DPF server logic for updating or not dpf vector.

**Proposed fixes**
- DPFVector are now always "committed" to dpf server whether they have been modified or not (as checking for modification is very slow)
- Performance is highly improved for server version checks by caching previous checks

**Benchmark**
Test is taken for #2201, changing params are:
- pydpf-core version
- server types: in process/gRPC
- vector operations: data is only "gotten", or data is "gotten" and "set"

| Server type | Vec operation |      0.13.6    |  0.13.7.dev |
|----------|:----------|:-------------:|------:|
| In Process | Get |  0.009 s | 0.008 s |
| In Process | Get/Set |  0.012 s |   0.011 s |
| gRPC | Get  |  6.608 s |   3.773 s |
| gRPC | Get/Set| 8.707 s |    3.707 s |